move sighting to a separate route to make it more scalable
