+++ src/config/routes.rb
@@ -10,10 +10,6 @@ Rails.application.routes.draw do
         scope '/:id' do
           get '/stops' => 'services#stops'
           get '/' => 'services#show'
-          scope :sighting do
-            get '/' => 'services#get_sighting'
-            post '/' => 'services#post_sighting'
-          end
         end
       end
 
@@ -34,6 +30,11 @@ Rails.application.routes.draw do
             put '/' => 'users#update_credit'
           end
         end
+      end
+
+      scope :sighting do
+        get '/' => 'sighting#get_sighting'
+        post '/' => 'sighting#post_sighting'
       end
     end
   end
